+++ src/app/pages/abm-manufacturados/abm-manufacturados.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Plato } from 'src/app/models/Plato';
 import { ArticulosService } from 'src/app/services/articulos.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-abm-manufacturados',
@@ -20,9 +21,30 @@ export class AbmManufacturadosComponent implements OnInit {
   }
 
   deletePlato(id:number){
-    this.articuloService.deleteManufacturado(id).subscribe((res)=>{
-      window.location.reload();
+    Swal.fire({
+      text: "¿Está seguro que desea eliminar?",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.articuloService.deleteManufacturado(id).subscribe((res)=>{
+          Swal.fire({
+            icon: 'success',
+            title: 'Se eliminó correctamente',
+            showConfirmButton: false,
+            timer: 1000
+          })
+          setTimeout(() => {
+            window.location.reload();
+          }, 1000);
+          
+        })
+      }
     })
+    
   }
 
 }
